requirements.txt
streamlit
pytesseract
Pillow
opencv-python-headless
pdfkit
numpy
fpdf
easyocr
python-docx
